Move & refactor test
Authored by: bashonly
